ai-image-detector-backend/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
scipy
scikit-learn
werkzeug
xgboost>=1.7.0
lightgbm>=3.3.0
PyWavelets>=1.4.0
scikit-image>=0.19.0
tqdm>=4.64.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
